refactor: 💡 judgmentErrorState move to function directory
