+++ src/pages/contact/index.page.tsx
@@ -7,6 +7,7 @@ import {
   useState,
 } from 'react';
 import { useRouter } from 'next/router';
+import { judgmentErrorState } from '@/pages/contact/logic/judgmentErrorState';
 
 export type InitialInput = {
   name: string;
@@ -18,7 +19,7 @@ export type InitialInput = {
   postContent: string;
 };
 
-type ErrorState = Partial<Record<keyof InitialInput, string>>;
+export type ErrorState = Partial<Record<keyof InitialInput, string>>;
 
 export default function Contact() {
   const [selected, setSelected] = useState<string[]>([]);
@@ -45,38 +46,7 @@ export default function Contact() {
   async function onSubmit(event: FormEvent<HTMLFormElement>) {
     event.preventDefault();
 
-    const newErrors: ErrorState = {};
-
-    // 必須項目チェック（空文字チェック）
-    for (const [key, value] of Object.entries(input)) {
-      if (!value.trim()) {
-        newErrors[key as keyof InitialInput] = 'これは必須項目でやんすねぇ';
-      }
-
-      // カタカナだけ特別なチェック（空でなくても）
-      if (key === 'katakana' && value && !/^[ァ-ン　]+$/.test(value)) {
-        newErrors.katakana = 'カタカナのみで入力してください';
-      }
-
-      // メール形式チェック
-      if (
-        key === 'mail' &&
-        value &&
-        !(/^[a-zA-Z0-9@.,]+$/.test(value) && /@/.test(value))
-      ) {
-        newErrors.mail = '英数字のみで入力して必ず@を使ってください';
-      }
-
-      // 電話番号チェック
-      if (key === 'phone' && value) {
-        if (!/^[0-9]+$/.test(value)) {
-          newErrors.phone = '半角数字のみで入力してください';
-        } else if (value.length >= 20) {
-          newErrors.phone =
-            '入力されている数字が多い可能性があります。いや、多いです。';
-        }
-      }
-    }
+    const newErrors = judgmentErrorState(input);
 
     // エラーがあるなら送信せず、エラーをセット
     if (Object.keys(newErrors).length > 0) {
